feat: --help

+++ fish/helpers/bak.fish
@@ -1,8 +1,10 @@
 function bak
-    if test (count $argv) -ne 1
-        echo "Usage: bak <file>"
-        return 1
+    set -l usage "Usage: bak <file>"
+    set -l desc "Creates a copy of the file for backup with datetime information."
+    if show_help "$usage" "$desc" $argv[1]
+        return
     end
+
     set original_file $argv[1]
     if not test -f $original_file
         echo "Error: File '$original_file' not found."
@@ -14,6 +16,13 @@ function bak
 end
 
 function rmbak
+    set -l usage "Usage: rmbak"
+    set -l desc "Removes all backup files (*.bak) from the current directory."
+
+    if show_help "$usage" "$desc" $argv[1]
+        return
+    end
+
     echo "Removing backup files (*.bak)..."
     rm *.bak
 end
