+++ fish/helpers/docker.fish
@@ -1,5 +1,13 @@
-function dockerclean --description "Removes all unused Docker data (containers, images, volumes, etc.) forcefully"
+function dockerclean
+    set -l usage "Usage: dockerclean"
+    set -l desc "Removes all unused Docker data (containers, images, networks, build cache, volumes) forcefully."
+
+    if show_help "$usage" "$desc" $argv[1]
+        return
+    end
+
     echo "Cleaning up all unused Docker data (containers, images, networks, build cache, volumes)..."
+
     docker system prune -af --volumes
     if test $status -ne 0
         echo "An error occurred during Docker system prune." >&2
@@ -8,7 +16,14 @@ function dockerclean --description "Removes all unused Docker data (containers, 
     echo "Docker system clean-up complete."
 end
 
-function dockerstop --description "Stops all running Docker containers"
+function dockerstop
+    set -l usage "Usage: dockerstop"
+    set -l desc "Stops all running Docker containers."
+
+    if show_help "$usage" "$desc" $argv[1]
+        return
+    end
+
     echo "Stopping all running Docker containers..."
     set running_containers (docker ps -q)
